Move ast queriable
