--- src/ast/expr/mod.rs
@@ -1,2 +1,0 @@
-pub mod query;
-
